[ASDisplayNode] Revise assertion to log until Issue #145 is addressed.
For apps that include pre-existing calls to reloadData, this assertion has made
development difficult on all recent versions of Texture.

Since triggering it can't be reliably used as signal by the app developer, and
halts a development session (sometimes immediately on app start), it should be
disabled until it is reliable again.

See issue #145 for details.