[Yoga] Delete YOGA_TREE_CONTIGOUS gating and permanently enable.
NOTE: Yoga support is still highly experimental and is not
planned to become a supported / documented mode of the framework.
We recommend that most apps use ASStackLayoutSpec and the other
specs: http://texturegroup.org/docs/layout2-layoutspec-types.html

This reduces complexity of the Yoga integration points, and is
a first step towards further minimization and formalization of
the Yoga footprint in the framework.