Update corner-rounding.md (#482)
* Update corner-rounding.md

If I understand right, both Precomposited corners techniques uses `[path clip]`, right? If so, I think it will be better to move it to the general description of the section, as currently it seems as only the second option uses `[path clip]`.

* Update corner-rounding.md

Improved positioning.

* Update corner-rounding.md

Syntax consistency: Changed `off-screen` to `offscreen`, as both syntaxes currently appears in the text.