Fix bug that breaks ASNodeController docs page
When Obj-C is selected on this page, the bottom half of the documentation is missing. When Swift is selected, malformed content appears where the documentation is nested inside the code block.

Repro'd here in Safari and Chrome: http://texturegroup.org/docs/containers-asnodecontroller.html

I think the bug is that <PhotoCellNode> is being parsed as an HTML tag, and this should fix it.