[Dev docs] Fix links of previous and next pages
Our pages are organized in 2 dirs: /docs/ and /development/. Within a dir, prevPage and nextPage are relative so the dir should be appended to the href link. The dir is available in permalink var. Some prevPage and nextPage point to a page in a different dir in which case they must not be relative. Instead they must include the dir and thus won't be modified.