[Dev docs] Fix previous and next page links (#1601)
Our pages are organized in 2 dirs: /docs/ and /development/.

Within a dir, prevPage and nextPage are relative (to the dir) so its path should be appended to their href links (see docs/_layouts/docs.html). The dir is available in the page's permalink var.

Some prevPage and nextPage point to a page in a different dir in which case they must not be relative. Instead they must include the other dir's path and thus their links won't be modified.

Also removed some pages that are no longer needed and fixed some prevPage and nextPage values.

I tested this diff locally and all links worked fine.