Set up GitHub Actions CI (#1623)
GitHub Actions now supports CI and it's free for open source projects!

Our current CI runs on a single machine and needs 1h 20m for each build. GitHub CI often finishes everything in less than 10m!

The setup is straight-forward. The only missing feature is the ability to exclude certain files so that if a commit only contains those files, the CI won't test it at all (see exclude-from-build.json). Given that GitHub CI is fast and is not a serial queue, I wouldn't mind testing such commit as well.

* Fix and optimize steps that build example projects

- examples-pt3 builds the same examples as examples-pt1 instead of building the remaining projects. So I fixed it.
- Add another mode (examples-pt4) to distribute the load with existing steps.
- Build extra examples (examples-extra-pt{1, 2, 3}) which weren't built before.
- Each examples-pt and examples-extra-pt builds 5 projects except examples-pt4 and examples-extra-pt3 which build 6.

* Use HTTPS URL for Weaver in Podfile of examples/ASDKgram

* Don't build extra examples for now